requirements.txt
pyyaml==6.0
markdown==3.4.1
openai==1.3.0
tweepy==4.14.0
